refactor: consolidate toolbar SVGs and unify clear icon
Remove duplicated toolbar SVG files from cad-viewer in favor of shared icons from cad-simple-viewer, replace clear-markups/measurements trash icons with a unified sweep icon, and reorder HTML export markup tools.
