+++ packages/cad-viewer/src/svg/index.ts
@@ -7,18 +7,13 @@ export { default as countlist } from './countlist.svg'
 export { default as defineAttribute } from './defineAttribute.svg'
 export { default as dynamicInput } from './dynamicInput.svg'
 export { default as editAttribute } from './editAttribute.svg'
-export { default as exportIcon } from './export.svg'
 export { default as fullScreen } from './fullScreen.svg'
-export { default as importIcon } from './import.svg'
 export { default as insertBlock } from './insertBlock.svg'
 export { default as language } from './language.svg'
-export { default as layer } from './layer.svg'
 export { default as line } from './line.svg'
 export { default as lineStyle } from './lineStyle.svg'
 export { default as lineWeight } from './lineWeight.svg'
 export { default as lineWidth } from './lineWidth.svg'
-export { default as markupPanel } from './markupPanel.svg'
-export { default as markupTools } from './markupTools.svg'
 export { default as mlightcad } from './mlightcad'
 export { default as mline } from './mline.svg'
 export { default as move } from './move.svg'
@@ -26,7 +21,6 @@ export { default as multiPoints } from './multiPoints.svg'
 export { default as offset } from './offset.svg'
 export { default as osnap } from './osnap.svg'
 export { default as orthoMode } from './orthoMode.svg'
-export { default as pan } from './pan.svg'
 export { default as polarTracking } from './polarTracking.svg'
 export { default as polygon } from './polygon.svg'
 export { default as polyline } from './polyline.svg'
@@ -34,23 +28,14 @@ export { default as properties } from './properties.svg'
 export { default as qselect } from './qselect.svg'
 export { default as ray } from './ray.svg'
 export { default as rect } from './rect.svg'
-export { default as revCircle } from './revCircle.svg'
-export { default as revCloud } from './revCloud.svg'
 export { default as revFreeDraw } from './revFreeDraw.svg'
-export { default as revRect } from './revRect.svg'
-export { default as revText } from './revText.svg'
-export { default as select } from './select.svg'
 export { default as setting } from './setting.svg'
-export { default as switchBg } from './switchBg.svg'
 export { default as xline } from './xline.svg'
-export { default as zoomToBox } from './zoomToBox.svg'
-export { default as zoomToExtent } from './zoomToExtent.svg'
 export * from './arc'
 export * from './circle'
 export * from './ellipse'
 export * from './hatch'
 export * from './layer'
 export * from './mtext'
-export * from './measure'
 export * from './pointstyle'
 export * from './spline'
