--- packages/cad-viewer/src/svg/measure/index.ts
@@ -1,7 +1,0 @@
-export { default as clearMeasurements } from './clearMeasurements.svg'
-export { default as measure } from './measure.svg'
-export { default as measureAngle } from './measureAngle.svg'
-export { default as measureArea } from './measureArea.svg'
-export { default as measureArc } from './measureArc.svg'
-export { default as measureDistance } from './measureDistance.svg'
-export { default as measurePoint } from './measurePoint.svg'
